Task 6 is done.
